Criação de contexto WebsiteWrapper, que centraliza o layout da aplicação.
